ft: refactor file structure, nest init command package code in config package
